full on 1/3 of the application
made it a working class
