+++ DataBase/task2/task2v1.py
@@ -1,0 +1,6 @@
+choice = input("What do you want to do (SELECT or INSERT): ")
+choice = choice.lower()
+
+if choice == "insert":
+    from insert_v_1 import INSERT
+    s = INSERT()
